+++ Tests/StreamChatUITests/SnapshotTests/ChatChannel/ChatChannelVC_Tests.swift
@@ -1615,6 +1615,41 @@ final class ChatChannelVC_Tests: XCTestCase {
 
         AssertSnapshot(vc, variants: [.defaultLight])
     }
+
+    func test_channelWithDraftMessage_whenDraftIsUpdatedFromEvent_whenThread_shouldNotUpdateChannelComposer() {
+        let draftMessage = DraftMessage.mock(text: "Draft Message")
+
+        let channel = ChatChannel.mock(cid: .unique, draftMessage: draftMessage)
+        channelControllerMock.channel_mock = channel
+
+        vc.view.layoutIfNeeded()
+
+        let updatedDraftMessage = DraftMessage.mock(threadId: .unique, text: "Updated draft")
+        channelControllerMock.mockCid = channel.cid
+        let event = DraftUpdatedEvent(
+            cid: channel.cid,
+            channel: channel,
+            draftMessage: updatedDraftMessage,
+            createdAt: .unique
+        )
+        vc.eventsController(vc.eventsController, didReceiveEvent: event)
+
+        AssertSnapshot(vc, variants: [.defaultLight])
+    }
+
+    func test_channelWithDraftMessage_whenDraftIsDeletedFromEvent_updatesDraftInComposer() {
+        let channel = ChatChannel.mock(cid: .unique, draftMessage: nil)
+        channelControllerMock.channel_mock = channel
+
+        vc.messageComposerVC.content.draftMessage(.mock(text: "Draft Message"))
+        XCTAssertFalse(vc.messageComposerVC.content.text.isEmpty)
+
+        channelControllerMock.mockCid = channel.cid
+        let event = DraftDeletedEvent(cid: channel.cid, threadId: nil, createdAt: .unique)
+        vc.eventsController(vc.eventsController, didReceiveEvent: event)
+
+        XCTAssertTrue(vc.messageComposerVC.content.text.isEmpty)
+    }
 }
 
 private extension ChatChannelVC_Tests {